requirements.txt
streamlit
sentence-transformers
faiss-cpu
numpy
tqdm
json5
protobuf
watchdog
